third_party/WebKit/LayoutTests/inspector/components/parsed-url-expected.txt

Tests inspector ParsedURL class

Parsing url: http://example.com/?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: 
  lastPathComponent: 
Parsing url: http://example.com/foo.html?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo.html
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: 
  lastPathComponent: foo.html
Parsing url: http://example.com/foo/bar.html?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo/bar.html
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: /foo
  lastPathComponent: bar.html
Parsing url: http://example.com/foo/bar/baz.html?queryParam1=value1&queryParam2=value2#fragmentWith/Many//Slashes
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo/bar/baz.html
  queryParams: queryParam1=value1&queryParam2=value2
  fragment: fragmentWith/Many//Slashes
  folderPathComponents: /foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com//?queryParam1=value1
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: //
  queryParams: queryParam1=value1
  fragment: undefined
  folderPathComponents: /
  lastPathComponent: 
Parsing url: http://example.com//foo.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: //foo.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /
  lastPathComponent: foo.html
Parsing url: http://example.com//foo/bar.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: //foo/bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo
  lastPathComponent: bar.html
Parsing url: http://example.com/foo//bar.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo//bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/
  lastPathComponent: bar.html
Parsing url: http://example.com//foo//bar.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: //foo//bar.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo/
  lastPathComponent: bar.html
Parsing url: http://example.com//foo/bar/baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: //foo/bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: //foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo//bar/baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo//bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo//bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo/bar//baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo/bar//baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/bar/
  lastPathComponent: baz.html
Parsing url: http://example.com///foo/bar/baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: ///foo/bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: ///foo/bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo////bar/baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo////bar/baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo////bar
  lastPathComponent: baz.html
Parsing url: http://example.com/foo/bar/////baz.html
  isValid: true
  scheme: http
  host: example.com
  port: undefined
  path: /foo/bar/////baz.html
  queryParams: 
  fragment: undefined
  folderPathComponents: /foo/bar////
  lastPathComponent: baz.html
